refactor(common): organize cost basis vocabulary
